basic refactoring
