+++ src/storageBackEnd/nativeBackEnd/nativehandler.py
@@ -1,0 +1,2 @@
+class NativeHandler:
+	pass
