+++ src/storageBackEnd/relationalBackEnd/sqlhandler.py
@@ -1,0 +1,2 @@
+class SqlHandler():
+    pass
