--- src/storageMS/relationalDB/src/sqlhandler.py
@@ -1,6 +1,0 @@
-
-
-
-class SqlHandler():
-    def __init__(self):
-        self.
